Rename files to remove pointless drm prefix
