+++ backend/meson.build
@@ -5,10 +5,10 @@ backend_files = files(
 	'session/session.c',
 	'drm/backend.c',
 	'drm/drm.c',
-	'drm/drm-atomic.c',
-	'drm/drm-legacy.c',
-	'drm/drm-properties.c',
-	'drm/drm-util.c',
+	'drm/iface_atomic.c',
+	'drm/iface_legacy.c',
+	'drm/properties.c',
+	'drm/util.c',
 	'libinput/backend.c',
 	'libinput/events.c',
 	'libinput/keyboard.c',
